Remove compiled files and library dependecies.
It is a good practice not to keep these files in the source code
repository. The compiled files can be generated locally any time, and
the dependencies, i.e. jar files, are going to be managed by Gradle,
which is going to download them if by the first time you need them.